+++ workon/cli.py
@@ -54,6 +54,22 @@ def _append_done_command(subparsers, parent):
     )
 
 
+def _append_open_command(subparsers, parent):
+    open_command = subparsers.add_parser(
+        'open', help='open an already started project',
+        formatter_class=argparse.ArgumentDefaultsHelpFormatter,
+        parents=[parent], add_help=False,
+    )
+
+    open_command.add_argument(
+        'project', help='project name under working directory to open')
+
+    open_command.add_argument(
+        '-e', '--editor', help='editor to use to open a project',
+        default=os.environ.get('WORKON_EDITOR')
+    )
+
+
 def _parse_args():
     parser = argparse.ArgumentParser()
 
@@ -74,6 +90,7 @@ def _parse_args():
 
     _append_start_command(subparsers, parent=parent_parser)
     _append_done_command(subparsers, parent=parent_parser)
+    _append_open_command(subparsers, parent=parent_parser)
 
     return parser.parse_args()