+++ workon/main.py
@@ -28,6 +28,8 @@ def main():
             script.start(args)
         elif args.command == 'done':
             script.done(args)
+        elif args.command == 'open':
+            script.open(args)
     except errors.ScriptError as exc:
         logging.error(exc)
         sys.exit(1)